Resolve circular import dependency and rename modules properly
