+++ main.py
@@ -2,3 +2,33 @@
 
 Copyright (c) 2021 Abdus Shaikh, Jason Wang, Samraj Aneja, Kevin Wang
 """
+
+
+class Main:
+    """Main runner for Prisoner's Dilemma.
+    """
+
+    def run_game(self) -> None:
+        """Run a game between two computer strategies.
+        """
+        # TODO: implement this function
+
+    def run_user_game(self) -> None:
+        """Run a game between a user and a computer strategy.
+        """
+        # TODO: implement this function
+
+    def run_tournament(self, show_heatmap: bool) -> None:
+        """Run a tournament between all strategies.
+
+        If <show_heatmap> is set, then display a heatmap that shows the match-ups
+        between the strategies.
+        """
+        # TODO: implement this function
+
+    def resolve_round(self, decision1: bool, decision2: bool) -> None:
+        """Takes Player 1's decision and Player 2's decision, determines
+        each player's points gained or lost, and mutates each Player's
+        curr_points to reflect it.
+        """
+        # TODO: implement this function
